move functions back to utils to avoid circular import
